+++ configs/eslint/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @quantum-design-configs/eslint
 
+## 2.0.0
+
+### Patch Changes
+
+- 适配依赖
 
 ## 2.0.0-beta.0