+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @quantum-design/utils
 
+## 2.0.0
+
+### Patch Changes
+
+- 适配依赖
+- Updated dependencies
+  - @quantum-design/shared@2.0.0
 
 ## 2.0.0-beta.0